Move files to more appropiate folders
