+++ app/(tabs)/home/index.tsx
@@ -1,4 +1,4 @@
-import ThemeButton from '@/app/components/ThemeButton';
+import ThemeButton from '@/src/components/ThemeButton';
 import { StyleSheet, Text } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
