+++ src/store.ts
@@ -1,7 +1,6 @@
 // store.ts
 import { create } from "zustand";
-import { createThemeSlice } from "./themeSlice";
-import { StoreState } from "./types";
+import { createThemeSlice, StoreState } from "./themeSlice";
 
 export const useThemeStore = create<StoreState>()((...a) => ({
   ...createThemeSlice(...a),
